【新增】供应商注销

+++ src/main/java/ses/service/sms/impl/SupplierServiceImpl.java
@@ -911,34 +911,40 @@ public class SupplierServiceImpl implements SupplierService {
         supplierAfterSaleDepMapper.deleteBySupplierId(supplierId);
        
         SupplierMatPro supplierMatPro = supplierMatProMapper.getMatProBySupplierId(supplierId);
-        supplierCertProMapper.deleteByPrimaryKey(supplierMatPro.getId());
+        if(supplierMatPro != null){
+        	supplierCertProMapper.deleteByPrimaryKey(supplierMatPro.getId());
+        }
         supplierMatProMapper.deleteBySupplierId(supplierId);
         
         SupplierMatSell matSell = supplierMatSellMapper.getMatSellBySupplierId(supplierId);
-        supplierCertSellMapper.deleteByPrimaryKey(matSell.getId());
+        if(matSell != null){
+        	supplierCertSellMapper.deleteByPrimaryKey(matSell.getId());
+        }
         supplierMatSellMapper.deleteByPrimaryKey(supplierId);
         
         SupplierMatEng matEng = supplierMatEngMapper.selectByPrimaryKey(supplierId);
-        supplierCertEngMapper.deleteByPrimaryKey(matEng.getId());
-        supplierRegPersonMapper.deleteByMatEngId(matEng.getId());
-        supplierAptituteMapper.deleteByPrimaryKey(matEng.getId());
+        if(matEng != null){
+        	supplierCertEngMapper.deleteByPrimaryKey(matEng.getId());
+            supplierRegPersonMapper.deleteByMatEngId(matEng.getId());
+            supplierAptituteMapper.deleteByPrimaryKey(matEng.getId());
+        }
         supplierMatEngMapper.deleteByPrimaryKey(supplierId);
         
         SupplierMatServe matServe = supplierMatServeMapper.getMatSeBySupplierId(supplierId);
-        supplierCertServeMapper.deleteByPrimaryKey(matServe.getId());
-        supplierMatServeMapper.deleteByPrimaryKey(matServe.getId());
-
+        if(matServe != null){
+	        supplierCertServeMapper.deleteByPrimaryKey(matServe.getId());
+	        supplierMatServeMapper.deleteByPrimaryKey(matServe.getId());
+        }
         supplierTypeRelateMapper.deleteBySupplierId(supplierId);
         List<SupplierItem> items= supplierItemMapper.getSupplierItem(supplierId);
-        
-        for(SupplierItem s:items){
-        	fileUploadMapper.deleteByBusinessId(s.getId());
+        if(!items.isEmpty()){
+	        for(SupplierItem s:items){
+	        	if(s.getId() != null){
+	        		fileUploadMapper.deleteByBusinessId(s.getId());
+	        	}
+	        }
         }
-        supplierItemMapper.deleteBySupplierId(supplierId);
-        
-        
-        
-        
+        supplierItemMapper.deleteBySupplierId(supplierId);  
     }
     
 }
